Refactor logo button
